Import as bids
- Move bids import logic into a new API / class
- Move common methods to ImportUtils class (maybe use inheritance
instead ?)
- Move tests
