+++ shanoir-ng-import/src/main/java/org/shanoir/ng/importer/ImporterApiController.java
@@ -18,26 +18,20 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
 import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
 import java.net.URLConnection;
 import java.nio.file.Files;
-import java.security.SecureRandom;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.zip.ZipOutputStream;
-
 import javax.validation.Valid;
 
 import org.apache.commons.io.FilenameUtils;
-import org.shanoir.ng.importer.dicom.DicomDirCreator;
 import org.shanoir.ng.exchange.imports.dicom.DicomDirGeneratorService;
 import org.shanoir.ng.exchange.model.ExExamination;
 import org.shanoir.ng.exchange.model.ExStudy;
@@ -46,7 +40,6 @@ import org.shanoir.ng.exchange.model.ExSubject;
 import org.shanoir.ng.exchange.model.Exchange;
 import org.shanoir.ng.importer.dicom.DicomDirToModelService;
 import org.shanoir.ng.importer.dicom.ImagesCreatorAndDicomFileAnalyzerService;
-import org.shanoir.ng.importer.dicom.ImportJobConstructorService;
 import org.shanoir.ng.importer.dicom.query.DicomQuery;
 import org.shanoir.ng.importer.dicom.query.QueryPACSService;
 import org.shanoir.ng.importer.eeg.brainvision.BrainVisionReader;
@@ -57,26 +50,19 @@ import org.shanoir.ng.importer.model.Channel;
 import org.shanoir.ng.importer.model.EegDataset;
 import org.shanoir.ng.importer.model.EegImportJob;
 import org.shanoir.ng.importer.model.Event;
-import org.shanoir.ng.importer.dto.ExaminationDTO;
-import org.shanoir.ng.importer.dto.StudyCardDTO;
 import org.shanoir.ng.importer.model.ImportJob;
 import org.shanoir.ng.importer.model.Patient;
 import org.shanoir.ng.importer.model.Serie;
 import org.shanoir.ng.importer.model.Study;
 import org.shanoir.ng.importer.model.Subject;
-import org.shanoir.ng.shared.configuration.RabbitMQConfiguration;
-import org.shanoir.ng.shared.core.model.IdName;
 import org.shanoir.ng.shared.exception.ErrorModel;
 import org.shanoir.ng.shared.exception.RestServiceException;
 import org.shanoir.ng.shared.exception.ShanoirException;
 import org.shanoir.ng.shared.exception.ShanoirImportException;
-import org.shanoir.ng.study.rights.StudyUser;
-import org.shanoir.ng.study.rights.StudyUserInterface;
 import org.shanoir.ng.utils.ImportUtils;
 import org.shanoir.ng.utils.KeycloakUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.ByteArrayResource;
@@ -94,9 +80,6 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import io.swagger.annotations.ApiParam;
 
 /**
@@ -124,21 +107,9 @@ public class ImporterApiController implements ImporterApi {
 
 	private static final Logger LOG = LoggerFactory.getLogger(ImporterApiController.class);
 
-	private static final SecureRandom RANDOM = new SecureRandom();
-
-	private static final String FILE_POINT = ".";
-
 	private static final String DICOMDIR = "DICOMDIR";
 
-	private static final String IMPORTJOB = "importJob.json";
-
-	private static final String APPLICATION_OCTET_STREAM = "application/octet-stream";
-
 	private static final String APPLICATION_ZIP = "application/zip";
-
-	private static final String UPLOAD_FILE_SUFFIX = ".upload";
-
-	private static final String ZIP_FILE_SUFFIX = ".zip";
 
 	/** The Constant KB. */
 	private static final int KB = 1024;
@@ -171,9 +142,6 @@ public class ImporterApiController implements ImporterApi {
 	private DicomDirToModelService dicomDirToModel;
 
 	@Autowired
-	private ImportJobConstructorService importJobConstructorService;
-
-	@Autowired
 	private ImagesCreatorAndDicomFileAnalyzerService imagesCreatorAndDicomFileAnalyzer;
 
 	@Autowired
@@ -181,9 +149,6 @@ public class ImporterApiController implements ImporterApi {
 
 	@Autowired
 	private QueryPACSService queryPACSService;
-
-	@Autowired
-	private RabbitTemplate rabbitTemplate;
 
 	public ResponseEntity<Void> uploadFiles(
 			@ApiParam(value = "file detail") @RequestPart("files") final MultipartFile[] files)
@@ -195,7 +160,7 @@ public class ImporterApiController implements ImporterApi {
 		try {
 			// not used currently
 			for (int i = 0; i < files.length; i++) {
-				saveTempFile(new File(importDir), files[i]);
+				ImportUtils.saveTempFile(new File(importDir), files[i]);
 			}
 			return new ResponseEntity<>(HttpStatus.OK);
 		} catch (IOException e) {
@@ -208,11 +173,11 @@ public class ImporterApiController implements ImporterApi {
 	public ResponseEntity<ImportJob> uploadDicomZipFile(
 			@ApiParam(value = "file detail") @RequestPart("file") final MultipartFile dicomZipFile)
 			throws RestServiceException {
-		if (dicomZipFile == null || !isZipFile(dicomZipFile)) {
+		if (dicomZipFile == null || !ImportUtils.isZipFile(dicomZipFile)) {
 			throw new RestServiceException(
 					new ErrorModel(HttpStatus.UNPROCESSABLE_ENTITY.value(), WRONG_CONTENT_FILE_UPLOAD, null));
 		}
-		if (!isZipFile(dicomZipFile)) {
+		if (!ImportUtils.isZipFile(dicomZipFile)) {
 			throw new RestServiceException(new ErrorModel(HttpStatus.UNPROCESSABLE_ENTITY.value(),
 					"Wrong content type of file upload, .zip required.", null));
 		}
@@ -222,8 +187,26 @@ public class ImporterApiController implements ImporterApi {
 			 * the import work folder (the root of everything): split imports to clearly
 			 * separate them into separate folders for each user
 			 */
-			File userImportDir = getUserImportDir();
-			File importJobDir = saveTempFileCreateFolderAndUnzip(userImportDir, dicomZipFile, true);
+			File userImportDir = ImportUtils.getUserImportDir(importDir);
+			
+			boolean createDicomDir = false;
+
+			File tempFile = ImportUtils.saveTempFile(userImportDir, dicomZipFile);
+
+			if (!ImportUtils.checkZipContainsFile(DICOMDIR, tempFile)) {
+				createDicomDir = true;
+			}
+			
+			File importJobDir = ImportUtils.saveTempFileCreateFolderAndUnzip(tempFile, dicomZipFile, true);
+
+			if (createDicomDir) {
+				LOG.info("DICOMDIR missing from zip file, generating one.");
+				final File dicomDir = new File(importJobDir, DICOMDIR);
+				if (!dicomDir.exists()) {
+					dicomDirGeneratorService.generateDicomDirFromDirectory(dicomDir, importJobDir);
+					LOG.info("DICOMDIR generated at path: {}", dicomDir.getAbsolutePath());
+				}
+			}
 
 			/**
 			 * 2. STEP: prepare patients list to be put into ImportJob: read DICOMDIR and
@@ -277,7 +260,7 @@ public class ImporterApiController implements ImporterApi {
 	public ResponseEntity<Void> startImportJob(
 			@ApiParam(value = "ImportJob", required = true) @Valid @RequestBody final ImportJob importJob)
 			throws RestServiceException {
-		File userImportDir = getUserImportDir();
+		File userImportDir = ImportUtils.getUserImportDir(importDir);
 		final Long userId = KeycloakUtil.getTokenUserId();
 		String tempDirId = importJob.getWorkFolder();
 		final File importJobDir = new File(userImportDir, tempDirId);
@@ -334,89 +317,6 @@ public class ImporterApiController implements ImporterApi {
 		return new ResponseEntity<>(importJob, HttpStatus.OK);
 	}
 
-	/**
-	 * This method takes a multipart file and stores it in a configured upload
-	 * directory in relation with the userId with a random name and the suffix
-	 * .upload
-	 *
-	 * @param file
-	 * @throws IOException
-	 */
-	private File saveTempFile(final File userImportDir, final MultipartFile file) throws IOException {
-		long n = createRandomLong();
-		File uploadFile = new File(userImportDir.getAbsolutePath(), Long.toString(n) + UPLOAD_FILE_SUFFIX);
-		file.transferTo(uploadFile);
-		return uploadFile;
-	}
-
-	/**
-	 * This method creates a random long number.
-	 * 
-	 * @return long: random number
-	 */
-	private long createRandomLong() {
-		long n = RANDOM.nextLong();
-		if (n == Long.MIN_VALUE) {
-			n = 0; // corner case
-		} else {
-			n = Math.abs(n);
-		}
-		return n;
-	}
-
-	/**
-	 * This method stores an uploaded zip file in a temporary file, creates a new
-	 * folder with the same name and unzips the content into this folder, and gives
-	 * back the folder with the content.
-	 * 
-	 * @param userImportDir
-	 * @param dicomZipFile
-	 * @return
-	 * @throws IOException
-	 * @throws RestServiceException
-	 */
-	private File saveTempFileCreateFolderAndUnzip(final File userImportDir, final MultipartFile dicomZipFile,
-			final boolean fromDicom) throws IOException, RestServiceException {
-		File tempFile = saveTempFile(userImportDir, dicomZipFile);
-		boolean createDicomDir = false;
-		if (fromDicom && !ImportUtils.checkZipContainsFile(DICOMDIR, tempFile)) {
-			createDicomDir = true;
-		}
-		String fileName = tempFile.getName();
-		int pos = fileName.lastIndexOf(FILE_POINT);
-		if (pos > 0) {
-			fileName = fileName.substring(0, pos);
-		}
-		File unzipFolderFile = new File(tempFile.getParentFile().getAbsolutePath() + File.separator + fileName);
-		if (!unzipFolderFile.exists()) {
-			unzipFolderFile.mkdirs();
-		} else {
-			throw new RestServiceException(new ErrorModel(HttpStatus.UNPROCESSABLE_ENTITY.value(),
-					"Error while unzipping file: folder already exists.", null));
-		}
-		ImportUtils.unzip(tempFile.getAbsolutePath(), unzipFolderFile.getAbsolutePath());
-		tempFile.delete();
-		if (createDicomDir) {
-			LOG.info("DICOMDIR missing from zip file, generating one.");
-			final File dicomDir = new File(unzipFolderFile, DICOMDIR);
-			if (!dicomDir.exists()) {
-				dicomDirGeneratorService.generateDicomDirFromDirectory(dicomDir, unzipFolderFile);
-				LOG.info("DICOMDIR generated at path: {}", dicomDir.getAbsolutePath());
-			}
-		}
-		return unzipFolderFile;
-	}
-
-	/**
-	 * Check if sent file is of type .zip.
-	 *
-	 * @param file
-	 */
-	private boolean isZipFile(final MultipartFile file) {
-		return file.getOriginalFilename().endsWith(ZIP_FILE_SUFFIX) || file.getContentType().equals(APPLICATION_ZIP)
-				|| file.getContentType().equals(APPLICATION_OCTET_STREAM);
-	}
-
 	@Override
 	public ResponseEntity<ImportJob> importDicomZipFile(
 			@ApiParam(value = "file detail") @RequestBody final String dicomZipFilename) throws RestServiceException {
@@ -442,57 +342,6 @@ public class ImporterApiController implements ImporterApi {
 		}
 	}
 
-
-	private boolean isZipFileFromFile(final File file) {
-		return file != null && file.getName().endsWith(ZIP_FILE_SUFFIX);
-	}
-
-	/**
-	 * This method stores an uploaded zip file in a temporary file, creates a new
-	 * folder with the same name and unzips the content into this folder, and gives
-	 * back the folder with the content.
-	 * 
-	 * @param userImportDir
-	 * @param dicomZipFile
-	 * @return
-	 * @throws IOException
-	 * @throws RestServiceException
-	 */
-	private File saveTempFileCreateFolderAndUnzipFromFile(final File userImportDir, final File dicomZipFile)
-			throws IOException, RestServiceException {
-		File tempFile = saveTempFileFromFile(userImportDir, dicomZipFile);
-		String fileName = tempFile.getName();
-		int pos = fileName.lastIndexOf(FILE_POINT);
-		if (pos > 0) {
-			fileName = fileName.substring(0, pos);
-		}
-		File unzipFolderFile = new File(tempFile.getParentFile().getAbsolutePath() + File.separator + fileName);
-		if (!unzipFolderFile.exists()) {
-			unzipFolderFile.mkdirs();
-		} else {
-			throw new RestServiceException(new ErrorModel(HttpStatus.UNPROCESSABLE_ENTITY.value(),
-					"Error while unzipping file: folder already exists.", null));
-		}
-		ImportUtils.unzip(tempFile.getAbsolutePath(), unzipFolderFile.getAbsolutePath());
-		tempFile.delete();
-		return unzipFolderFile;
-	}
-
-	/**
-	 * This method takes a multipart file and stores it in a configured upload
-	 * directory in relation with the userId with a random name and the suffix
-	 * .upload
-	 *
-	 * @param file
-	 * @throws IOException
-	 */
-	private File saveTempFileFromFile(final File userImportDir, final File file) throws IOException {
-		long n = createRandomLong();
-		File uploadFile = new File(userImportDir.getAbsolutePath(), Long.toString(n) + UPLOAD_FILE_SUFFIX);
-		Files.move(file.toPath(), uploadFile.toPath());
-		return uploadFile;
-	}
-
 	@Override
 	/**
 	 * This method load an EEG file, unzip it and load an import job with the
@@ -507,7 +356,7 @@ public class ImporterApiController implements ImporterApi {
 				throw new RestServiceException(
 						new ErrorModel(HttpStatus.UNPROCESSABLE_ENTITY.value(), NO_FILE_UPLOADED, null));
 			}
-			if (!isZipFile(eegFile)) {
+			if (!ImportUtils.isZipFile(eegFile)) {
 				throw new RestServiceException(
 						new ErrorModel(HttpStatus.UNPROCESSABLE_ENTITY.value(), WRONG_CONTENT_FILE_UPLOAD, null));
 			}
@@ -524,7 +373,9 @@ public class ImporterApiController implements ImporterApi {
 			}
 
 			// Unzip the file and get the elements
-			File importJobDir = saveTempFileCreateFolderAndUnzip(userImportDir, eegFile, false);
+			File tempFile = ImportUtils.saveTempFile(userImportDir, eegFile);
+
+			File importJobDir = ImportUtils.saveTempFileCreateFolderAndUnzip(tempFile, eegFile, false);
 
 			EegImportJob importJob = new EegImportJob();
 			importJob.setWorkFolder(importJobDir.getName());
@@ -735,8 +586,8 @@ public class ImporterApiController implements ImporterApi {
 
 	@Override
 	public ResponseEntity<String> createTempDir() throws RestServiceException {
-		final File userImportDir = getUserImportDir();
-		long n = createRandomLong();
+		final File userImportDir = ImportUtils.getUserImportDir(importDir);
+		long n = ImportUtils.createRandomLong();
 		File tempDirForImport = new File(userImportDir, Long.toString(n));
 		if (!tempDirForImport.exists()) {
 			tempDirForImport.mkdirs();
@@ -747,20 +598,10 @@ public class ImporterApiController implements ImporterApi {
 		return new ResponseEntity<String>(tempDirForImport.getName(), HttpStatus.OK);
 	}
 
-	private File getUserImportDir() {
-		final Long userId = KeycloakUtil.getTokenUserId();
-		final String userImportDirFilePath = importDir + File.separator + Long.toString(userId);
-		final File userImportDir = new File(userImportDirFilePath);
-		if (!userImportDir.exists()) {
-			userImportDir.mkdirs(); // create if not yet existing
-		} // else is wanted case, user has already its import directory
-		return userImportDir;
-	}
-
 	@Override
 	public ResponseEntity<Void> uploadFile(@PathVariable("tempDirId") String tempDirId,
 			@RequestParam("file") MultipartFile file) throws RestServiceException, IOException {
-		final File userImportDir = getUserImportDir();
+		final File userImportDir = ImportUtils.getUserImportDir(importDir);
 		final File importJobDir = new File(userImportDir, tempDirId);
 		// only continue in case of existing temp dir id
 		if (importJobDir.exists()) {
@@ -783,7 +624,7 @@ public class ImporterApiController implements ImporterApi {
 	public ResponseEntity<Void> startImport(@RequestBody Exchange exchange)
 			throws RestServiceException, FileNotFoundException, IOException {
 		// 1. Check if uploaded data are complete (to be done a little later)
-		final File userImportDir = getUserImportDir();
+		final File userImportDir = ImportUtils.getUserImportDir(importDir);
 		final File tempDir = new File(userImportDir, exchange.getTempDirId());
 
 		final File dicomDir = new File(tempDir, DICOMDIR);
@@ -841,223 +682,6 @@ public class ImporterApiController implements ImporterApi {
 	}
 
 	/**
-	 * This methods import a bunch of datasets from a Shanoir Exchange Format (based
-	 * on BIDS format)
-	 * 
-	 * @param bidsFile
-	 *            the file
-	 * @throws ShanoirException
-	 *             when something gets wrong during the import
-	 * @throws IOException
-	 *             when IO fails
-	 * @throws RestServiceException
-	 */
-	@Override
-	public ResponseEntity<ImportJob> importAsBids(
-			@ApiParam(value = "file detail") @RequestPart("file") final MultipartFile bidsFile)
-			throws RestServiceException, ShanoirException, IOException {
-		// Check that the file is not null and well zipped
-		if (bidsFile == null) {
-			throw new RestServiceException(
-					new ErrorModel(HttpStatus.UNPROCESSABLE_ENTITY.value(), NO_FILE_UPLOADED, null));
-		}
-		if (!isZipFile(bidsFile)) {
-			// .SEF ?
-			throw new RestServiceException(
-					new ErrorModel(HttpStatus.UNPROCESSABLE_ENTITY.value(), WRONG_CONTENT_FILE_UPLOAD, null));
-		}
-		// Todo: what if we are coming from SHUP ?
-
-		// Create tmp folder and unzip archive
-		final File userImportDir = getUserImportDir();
-		File importJobDir = saveTempFileCreateFolderAndUnzip(userImportDir, bidsFile, false);
-		// Deserialize participants.tsv => Do a call to studies API to create
-		// corresponding subjects
-		File participantsFile = new File(importJobDir.getAbsolutePath() + "/participants.tsv");
-		if (!participantsFile.exists()) {
-			throw new ShanoirException("participants.tsv file is mandatory");
-		}
-
-		ObjectMapper mapper = new ObjectMapper();
-
-		SimpleModule module = new SimpleModule();
-		module.addAbstractTypeMapping(StudyUserInterface.class, StudyUser.class);
-		mapper.registerModule(module);
-		// Here we wait for the response => to be sure that the subjects are created
-		String participantString = (String) rabbitTemplate.convertSendAndReceive(RabbitMQConfiguration.SUBJECTS_QUEUE, participantsFile.getAbsolutePath());
-		List<IdName> participants = Arrays.asList(mapper.readValue(participantString, IdName[].class));
-		// If we receive a unique subject with no ID => It's an error
-		if (participants.size() == 1 && participants.get(0).getId() == null) {
-			throw new ShanoirException(participants.get(0).getName());
-		}
-
-		File studyDescriptionFile = new File(importJobDir.getAbsolutePath() + "/dataset_description.json");
-		if (!studyDescriptionFile.exists()) {
-			throw new ShanoirException("studyDescriptionFile file is mandatory");
-		}
-
-		// Then import data
-		File sourceData = new File(importJobDir.getAbsolutePath() + "/sourcedata");
-		if (!sourceData.exists()) {
-			throw new ShanoirException("sourcedata folder is mandatory");
-		}
-
-		// 2) Import Datasets
-		File[] subjectFiles = sourceData.listFiles(new FilenameFilter() {
-			@Override
-			public boolean accept(File dir, String name) {
-				return name.startsWith("sub-");
-			}
-		});
-		ImportJob job = null;
-
-		for (File subjFile : subjectFiles) {
-			// Get subjectName
-			String subjectName = subjFile.getName().substring("sub-".length());
-
-			// Read shanoirImportFile
-			File shanoirImportFile = new File(subjFile.getAbsolutePath() + "/shanoir-import.json");
-
-			if (!shanoirImportFile.exists()) {
-				throw new ShanoirException("shanoir-import.json file is mandatory in subject folder");
-			}
-
-			ObjectMapper objectMapper = new ObjectMapper();
-			ImportJob sid = objectMapper.readValue(shanoirImportFile, ImportJob.class);
-
-			String studyCardAsString = (String) rabbitTemplate.convertSendAndReceive(RabbitMQConfiguration.FIND_STUDY_CARD_QUEUE, sid.getStudyCardId());
-
-			if (studyCardAsString == null) {
-				throw new ShanoirException(
-						"StudyCard with ID " + sid.getStudyCardId() + " does not exists.");
-			}
-
-			StudyCardDTO studyCard = objectMapper.readValue(studyCardAsString, StudyCardDTO.class);
-
-
-			if (studyCard.getStudyId() != sid.getStudyId()) {
-				throw new ShanoirException("Study with ID " + sid.getStudyId() + " does not exists.");
-			}
-			if (studyCard.isDisabled()) {
-				throw new ShanoirException("StudyCard with ID " + sid.getStudyCardId() + " is currently disabled, please select another one.");
-			}
-			
-			// Subject based on folder name
-			Long subjectId = getSubjectIdByName(subjectName, participants);
-			if (subjectId == null) {
-				throw new ShanoirException(
-						"Subject " + subjectName + " could not be created. Please check participants.tsv file.");
-			}
-
-			// If there is no DICOMDIR: create it
-			File dicomDir = new File(subjFile.getAbsolutePath() + "/DICOM/DICOMDIR");
-			if (!dicomDir.exists()) {
-				DicomDirCreator creator = new DicomDirCreator(subjFile.getAbsolutePath() + "/DICOMDIR",
-						subjFile.getAbsolutePath() + "/DICOM");
-				creator.start();
-			}
-
-			// Zip data folders to be able to call ImporterAPIController.uploadDicomZipFile
-			FileOutputStream fos = new FileOutputStream(subjFile.getAbsolutePath() + ".zip");
-			ZipOutputStream zipOut = new ZipOutputStream(fos);
-
-			ImportUtils.zipFile(subjFile, subjFile.getName(), zipOut, true);
-
-			zipOut.close();
-			fos.close();
-			MockMultipartFile multiPartFile = new MockMultipartFile(subjFile.getName(), subjFile.getName() + ".zip",
-					APPLICATION_ZIP, new FileInputStream(subjFile.getAbsolutePath() + ".zip"));
-
-			// Send data folder to import API and get import job
-			ResponseEntity<ImportJob> entity = this.uploadDicomZipFile(multiPartFile);
-
-			// Complete ImportJob to use startImportJob
-			job = entity.getBody();
-
-			// Construire l'arborescence
-			job.setAcquisitionEquipmentId(studyCard.getAcquisitionEquipmentId());
-			job.setStudyId(sid.getStudyId());
-
-			job.setFromPacs(false);
-			job.setFromShanoirUploader(false);
-			job.setFromDicomZip(true);
-			job.setConverterId(studyCard.getNiftiConverterId());
-
-			for (Patient pat : job.getPatients()) {
-				pat.setPatientName(subjectName);
-				Subject subject = new Subject();
-				subject.setId(subjectId);
-				subject.setName(subjectName);
-				pat.setSubject(subject);
-
-				// Select all series to be imported
-				for (Study study : pat.getStudies()) {
-					for (Serie serie : study.getSeries()) {
-						serie.setSelected(Boolean.TRUE);
-					}
-				}
-			}
-
-			// Create a new examination if not existing
-			if (sid.getExaminationId() == null || sid.getExaminationId().equals(Long.valueOf(0l))) {
-				// Create examination => We actually need its ID so do a direct API call
-				
-				// Get center ID
-				String centerAsString = (String) rabbitTemplate.convertSendAndReceive(RabbitMQConfiguration.ACQUISITION_EQUIPEMENT_CENTER_QUEUE, studyCard.getAcquisitionEquipmentId());
-				IdName center = objectMapper.readValue(centerAsString, IdName.class);
-
-				ExaminationDTO examDTO = new ExaminationDTO();
-				// Construct DTO
-				// get center from study card => equipment => center
-				examDTO.setCenter(new IdName(center.getId(), center.getName()));
-				examDTO.setPreclinical(false); // Pour le moment on fait que du DICOM
-				examDTO.setStudy(new IdName(sid.getStudyId(), sid.getStudyName()));
-				examDTO.setSubject(new IdName(subjectId, subjectName));
-				examDTO.setExaminationDate(job.getPatients().get(0).getStudies().get(0).getStudyDate());
-				examDTO.setComment(job.getPatients().get(0).getStudies().get(0).getStudyDescription());
-				
-				mapper.registerModule(new JavaTimeModule());
-				String examAsString = (String) rabbitTemplate.convertSendAndReceive(RabbitMQConfiguration.EXAMINATION_CREATION_QUEUE, mapper.writeValueAsString(examDTO));
-
-				examDTO = mapper.readValue(examAsString, ExaminationDTO.class);
-				job.setExaminationId(examDTO.getId());
-			}
-
-			// Next API call => StartImportJob
-			ResponseEntity<Void> result = this.startImportJob(job);
-			if (!result.getStatusCode().equals(HttpStatus.OK)) {
-				throw new ShanoirException("Error while importing subject: " + subjectName);
-			}
-		}
-
-		// TODO ONE DAY: Copy "other" files to the bids folder
-		// Copy non datasets elements
-		// Don't copy "data" folder
-		// Don't copy examination_description.json
-		// copy /sourceData??, /code and / files (readme, changes, participants.tsv,
-		// participants.json, etc..)
-		return new ResponseEntity<>(job, HttpStatus.OK);
-	}
-
-	/**
-	 * Get the ID of a subject from its name and a list of subject
-	 * 
-	 * @param name
-	 *            the name of the subject to find
-	 * @param subjects
-	 *            the list of subjects to supply
-	 * @return the ID of the subject corresponding to the name, null otherwise
-	 */
-	public Long getSubjectIdByName(String name, List<IdName> subjects) {
-		for (IdName sub : subjects) {
-			if (sub.getName().equals(name)) {
-				return sub.getId();
-			}
-		}
-		return null;
-	}
-
-	/**
 	 * This methods returns a dicom file
 	 * 
 	 * @param path
@@ -1072,7 +696,7 @@ public class ImporterApiController implements ImporterApi {
 	public ResponseEntity<ByteArrayResource> getDicomImage(@ApiParam(value = "path", required=true)  @RequestParam(value = "path", required = true) String path)
 			throws RestServiceException, IOException {
 
-		final File userImportDir = getUserImportDir();
+		final File userImportDir = ImportUtils.getUserImportDir(importDir);
 		String pathInfo = userImportDir.getAbsolutePath() + File.separator + path;
 		URL url = new URL("file:///" + pathInfo);
 		final URLConnection uCon = url.openConnection();
